FIX: avoid auto unboxing
git-svn-id: https://cvs.dai-labor.de/jiac/jiac-tng/trunk@41275 d8bac29e-b7f5-0310-9bdb-a06bcefdf3b8

+++ agentCoreAPI/src/main/java/de/dailab/jiactng/agentcore/action/Session.java
@@ -439,7 +439,7 @@ public class Session implements IFact {
    private void readObject(ObjectInputStream ois) throws ClassNotFoundException, IOException {
      ois.defaultReadObject();
      Long timePassed = (Long)ois.readObject();
-     creationTime = Long.valueOf(System.currentTimeMillis() - timePassed);
+     creationTime = Long.valueOf(System.currentTimeMillis() - timePassed.longValue());
    }
